Middleware en invocable
