+++ Tools/Middlewares/DBConnector.php
@@ -1,0 +1,42 @@
+<?php declare(strict_types = 1);
+namespace LibertAPI\Tools\Middlewares;
+
+use Psr\Http\Message\ServerRequestInterface as IRequest;
+use Psr\Http\Message\ResponseInterface as IResponse;
+use Doctrine\DBAL;
+
+/**
+ * Connexion DB
+ *
+ * @since 1.0
+ */
+final class DBConnector extends \LibertAPI\Tools\AMiddleware
+{
+    public function __invoke(IRequest $request, IResponse $response, callable $next) : IResponse
+    {
+        try {
+            $configuration = json_decode(file_get_contents(ROOT_PATH . 'configuration.json'));
+            $dbh = new \PDO(
+                'mysql:host=' . $configuration->db->serveur . ';dbname=' . $configuration->db->base,
+                $configuration->db->utilisateur,
+                $configuration->db->mot_de_passe,
+                [\PDO::MYSQL_ATTR_INIT_COMMAND => 'SET NAMES \'utf8\';']
+            );
+            $connexion = DBAL\DriverManager::getConnection([
+                    'pdo' => $dbh
+                ]
+            );
+            $this->getApp()->getContainer()['storageConnector'] = $connexion;
+
+            return $next($request, $response);
+        /* Fallback */
+        } catch (\Exception $e) {
+            return call_user_func(
+                $this->getApp()->getContainer()->errorHandler,
+                $request,
+                $response,
+                $e
+            );
+        }
+    }
+}
