+++ Tools/Middlewares/Identificator.php
@@ -1,0 +1,80 @@
+<?php declare(strict_types = 1);
+namespace LibertAPI\Tools\Middlewares;
+
+use Psr\Http\Message\ServerRequestInterface as IRequest;
+use Psr\Http\Message\ResponseInterface as IResponse;
+use \LibertAPI\Tools\Libraries\AEntite;
+use \LibertAPI\Tools\Libraries\ARepository;
+use \LibertAPI\Tools\Helpers\Formatter;
+use \LibertAPI\Utilisateur;
+
+/**
+ * Réalise l'identification. En profite pour pinger.
+ *
+ * @since 1.0
+ */
+final class Identificator extends \LibertAPI\Tools\AMiddleware
+{
+    public function __invoke(IRequest $request, IResponse $response, callable $next) : IResponse
+    {
+        $container = $this->getApp()->getContainer();
+        $repoUtilisateur = new Utilisateur\UtilisateurRepository(
+            new Utilisateur\UtilisateurDao($container->storageConnector)
+        );
+        // @TODO: doublon de middleware, à changer.
+        $openRoutes = ['Authentification', 'HelloWorld'];
+        $ressourcePath = $request->getAttribute('nomRessources');
+        if (in_array($ressourcePath, $openRoutes, true)) {
+            return $next($request, $response);
+        } elseif ($this->isIdentificationOK($request, $repoUtilisateur)) {
+             // Ping de last_access
+            $repoUtilisateur->updateDateLastAccess($this->utilisateur);
+
+            $container['currentUser'] = $this->utilisateur;
+            return $next($request, $response);
+        }
+
+        return call_user_func(
+            $container->unauthorizedHandler,
+            $request,
+            $response
+        );
+    }
+
+    private function isIdentificationOK(IRequest $request, ARepository $repository) : bool
+    {
+        $token = $request->getHeaderLine('Token');
+        if (empty($token)) {
+            return false;
+        }
+        try {
+            $this->utilisateur = $repository->find([
+                'token' => $token,
+                'gt_date_last_access' => $this->getDateLastAccessAuthorized()
+            ]);
+            return $this->utilisateur instanceof AEntite;
+        } catch (\UnexpectedValueException $e) {
+            return false;
+        }
+    }
+
+    /**
+     * Retourne la date limite de dernier accès pour être considéré en ligne
+     *
+     * @return string
+     */
+    private function getDateLastAccessAuthorized() : string
+    {
+        return Formatter::timeToSQLDatetime(time() - static::DUREE_SESSION);
+    }
+
+    /**
+     * @var AEntite | null
+     */
+    private $utilisateur;
+
+    /**
+     * @var int Durée de validité du token fourni, en secondes
+     */
+    const DUREE_SESSION = 30*60;
+}
